Allow structured, post-composed phenotype annotations.
This commit amends the PhenotypeAnnotation class by adding two slots:
`affected_entity` and `qualifier`.

`affected_entity` represents any entity that is affected by the
phenotype that is being described. It can be an anatomical structure, a
developmental stage, or a biological process. The slot has a dual
purpose:

1. It can refine a phenotype that is otherwise described in
   `phenotype_terms`, by indicating which organ, stage, or process is
   affected by a phenotype.

2. It can describe a phenotype on its own (with `phenotype_terms` left
   empty), if the main characteristic of the phenotype is that it
   affects a particular organ, stage, or process (this is especially the
   case for some FlyBase-style phenotype annotations).

`qualifier` is used to further refine ("qualify") a phenotype that is
otherwise described either by `phenotype_terms` or by `affected_entity`.
It must contain at the very least an ontological term from either PATO
or one of the phenotype ontologies (for the phenotype ontologies that
have "modifier" terms). It may also contain a `related_entity`, which
describes an additional entity (in the same sense as `affected_entity`)
that is linked to the phenotype through the qualifier term. This slot is
especially intended to allow representing ZFIN-style "Entity+Qualifier"
(EQ) phenotype statements.