Re-type BioSampleAge as a AuditedObject.
BioSampleAge is currently classified as a SubmittedObject. This means
it is supposed to represent a public entity that has its own CURIE (and
also means that, when submitting data, we must include a
`data_provider_dto`).

This should not be necessary for something as simple as the "age" of a
biological sample.

This is also inconsistent with the way the
`htp_expression_sample_location` slot is defined (that slot expects a
`AnatomicalSite`, which is a `AuditedObject`, not a `SubmittedObject`).
This means that the _age_ of a sample is expected to be an identifiable
entity, but not its _location_.

Here we remove the inconsistency by making `BioSampleAge` a
`AuditedObject` (and `BioSampleAgeDTO` a `AuditedObjectDTO`).