Fix requirements on the BioSampleGenomicInformationDTO class.
The `BioSampleGenomicInformationDTO` class reuses the
`allele_identifier` and `agm_identifier` slots from the `allele.yaml`
module. These slots are defined, in that module, as being *required*.

Since we do _not_ want them to be unconditionally required in the
context of the `BioSampleGenomicInformationDTO` class (instead we want
to require only _one_ of them, using a class-specific rule), we must
explicitly override their default `required: true` value to mark them as
being not required in that class.